refactor(auth): extract shared base64url helper (#163)
